refactor: mv queryButtons to QueryActions
